backend/requirements.txt
flask
flask-cors
flask-swagger-ui
python-dotenv
requests